refactor pullData controller action to lean out the function
